+++ src/app/Http/Controllers/Auth/RegisterController.php
@@ -3,11 +3,20 @@
 namespace App\Http\Controllers\Auth;
 
 use App\Http\Controllers\Controller;
+use App\Models\GroupUserInvitation;
 use App\Providers\RouteServiceProvider;
 use App\Models\User;
+use Arr;
+use Exception;
+use Illuminate\Auth\Events\Registered;
 use Illuminate\Foundation\Auth\RegistersUsers;
+use Illuminate\Http\JsonResponse;
+use Illuminate\Http\RedirectResponse;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Validator;
+use Illuminate\Http\Request;
+use Illuminate\Validation\Rule;
+use Illuminate\Validation\ValidationException;
 use Inertia\Inertia;
 use Inertia\Response;
 
@@ -29,11 +38,45 @@ class RegisterController extends Controller
     }
 
     /**
+     * @param Request $request
      * @return Response
      */
-    public function showRegistrationForm()
+    public function showRegistrationForm(Request $request)
     {
-        return Inertia::render('auth/register');
+        return Inertia::render('auth/register', [
+            'invitation' => [
+                'code' => $request->input('invitationCode'),
+                'email' => $request->input('email'),
+            ],
+        ]);
+    }
+
+    /**
+     * Handle a registration request for the application.
+     *
+     * @param Request $request
+     * @return RedirectResponse|JsonResponse
+     * @throws ValidationException
+     */
+    public function register(Request $request)
+    {
+        $this->validator($request->all())->validate();
+        $user = $this->create($request->all());
+
+        if ($request->input('invitation_code')) {
+            $user->markEmailAsVerified();
+        }
+        event(new Registered($user));
+
+        $this->guard()->login($user);
+
+        if ($response = $this->registered($request, $user)) {
+            return $response;
+        }
+
+        return $request->wantsJson()
+            ? new JsonResponse([], 201)
+            : redirect($this->redirectPath());
     }
 
     /**
@@ -42,21 +85,45 @@ class RegisterController extends Controller
      */
     protected function validator(array $data)
     {
-        return Validator::make($data, [
-            'first_name' => ['required', 'string', 'max:255'],
-            'last_name' => ['required', 'string', 'max:255'],
-            'email' => [
-                'required',
-                'string',
-                'email',
-                'max:255',
-                'unique:users',
+        return Validator::make(
+            $data,
+            [
+                'first_name' => ['required', 'string', 'max:255'],
+                'last_name' => ['required', 'string', 'max:255'],
+                'email' => [
+                    'required',
+                    'string',
+                    'email',
+                    'max:255',
+                    'unique:users',
+                ],
+                'company' => ['required', 'string', 'max:255'],
+                'password' => ['required', 'string', 'min:8', 'confirmed'],
+                'password_confirmation' => ['required', 'string', 'min:8'],
+                'terms' => ['required'],
+                'invitation_code' => [
+                    'nullable',
+                    Rule::requiredIf(env('INVITATION_ON_REGISTER', false)),
+                    Rule::exists(
+                        'group_user_invitations',
+                        'invitation_code'
+                    )->where(function ($query) use ($data) {
+                        return $query
+                            ->where('email', Arr::get($data, 'email'))
+                            ->where(
+                                'expired_at',
+                                '>',
+                                now()->toDateTimeString()
+                            );
+                    }),
+                ],
             ],
-            'company' => ['required', 'string', 'max:255'],
-            'password' => ['required', 'string', 'min:8', 'confirmed'],
-            'password_confirmation' => ['required', 'string', 'min:8'],
-            'terms' => ['required'],
-        ]);
+            [
+                'invitation_code.exists' => __(
+                    'Email address and Invitation code are mismatched.'
+                ),
+            ]
+        );
     }
 
     /**
@@ -73,5 +140,29 @@ class RegisterController extends Controller
             'password' => Hash::make($data['password']),
             'role' => User::ROLE_USER,
         ]);
+    }
+
+    /**
+     * The user has been registered.
+     *
+     * @param Request $request
+     * @param mixed $user
+     * @return mixed
+     * @throws Exception
+     */
+    protected function registered(Request $request, $user)
+    {
+        $invitations = GroupUserInvitation::where('email', $user->email)->get();
+
+        foreach ($invitations as $invitation) {
+            /** @var GroupUserInvitation $invitation */
+            if ($invitation->isActive()) {
+                $groupUsers = $invitation->group->users();
+                $groupUsers->attach($user->id, [
+                    'admin' => !$groupUsers->exists(),
+                ]);
+            }
+            $invitation->delete();
+        }
     }
 }