+++ src/resources/js/layouts/groups/main.vue
@@ -50,6 +50,24 @@
                                         Members
                                     </inertia-link>
                                 </li>
+                                <li class="nav-item" v-if="group.can.admin">
+                                    <inertia-link
+                                        :href="
+                                            route(
+                                                'groups.user-invitations.index',
+                                                group.id
+                                            )
+                                        "
+                                        class="nav-link rounded-0"
+                                        v-bind:class="{
+                                            active: route().current(
+                                                'groups.user-invitations.*'
+                                            ),
+                                        }"
+                                    >
+                                        Invitations
+                                    </inertia-link>
+                                </li>
                                 <li class="nav-item">
                                     <inertia-link
                                         :href="